Move visiblePagesContains method to PDFView extension
